// src/app/model/emigoLogin.ts
/**
 * Digital Archive Emigo App Server API
 * An interface to a emigo app of distributed nodes for personal storage. emigo-id - hex encoded sha256 of an emigo 4096 public key emigo-key - hex encoded public key
 *
 * OpenAPI spec version: 1.0.4
 * Contact: rosborne@coredb.org
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { NodeConnection } from './nodeConnection';


export interface EmigoLogin { 
    token: string;
    account: NodeConnection;
    service: NodeConnection;
}